+++ src/django_elasticsearch_dsl_drf/filter_backends/filtering/nested.py
@@ -2,7 +2,7 @@
 Nested filtering backend.
 """
 
-from elasticsearch_dsl.query import Q
+from elasticsearch.dsl.query import Q
 from django.core.exceptions import ImproperlyConfigured
 from django_elasticsearch_dsl import fields