+++ src/django_elasticsearch_dsl_drf/filter_backends/highlight.py
@@ -115,10 +115,10 @@ class HighlightBackend(BaseFilterBackend):
         :param queryset: Base queryset.
         :param view: View.
         :type request: rest_framework.request.Request
-        :type queryset: elasticsearch_dsl.search.Search
+        :type queryset: elasticsearch.dsl.search.Search
         :type view: rest_framework.viewsets.ReadOnlyModelViewSet
         :return: Updated queryset.
-        :rtype: elasticsearch_dsl.search.Search
+        :rtype: elasticsearch.dsl.search.Search
         """
         highlight_query_params = self.get_highlight_query_params(request)
         highlight_fields = self.prepare_highlight_fields(view)