+++ src/django_elasticsearch_dsl_drf/filter_backends/search/query_backends/match.py
@@ -1,4 +1,4 @@
-from elasticsearch_dsl.query import Q
+from elasticsearch.dsl.query import Q
 
 from .base import BaseSearchQueryBackend