+++ src/django_elasticsearch_dsl_drf/filter_backends/search/query_backends/multi_match.py
@@ -1,6 +1,6 @@
 import copy
 
-from elasticsearch_dsl.query import Q
+from elasticsearch.dsl.query import Q
 
 from .base import BaseSearchQueryBackend