+++ src/django_elasticsearch_dsl_drf/filter_backends/search/query_backends/nested.py
@@ -1,7 +1,7 @@
 import operator
 import six
 
-from elasticsearch_dsl.query import Q
+from elasticsearch.dsl.query import Q
 
 from .base import BaseSearchQueryBackend