+++ src/django_elasticsearch_dsl_drf/versions.py
@@ -26,7 +26,7 @@ def get_elasticsearch_version(default=(2, 0, 0)):
     :rtype: list
     """
     try:
-        from elasticsearch_dsl import __version__
+        from elasticsearch.dsl import __version__
         return __version__
     except ImportError:
         return default