+++ src/django_elasticsearch_dsl_drf/viewsets.py
@@ -10,9 +10,9 @@ import copy
 from django.http import Http404
 from django.core.exceptions import ImproperlyConfigured
 
-from elasticsearch_dsl import Search
-from elasticsearch_dsl.connections import connections
-from elasticsearch_dsl.query import MoreLikeThis
+from elasticsearch.dsl import Search
+from elasticsearch.dsl.connections import connections
+from elasticsearch.dsl.query import MoreLikeThis
 
 from rest_framework import status
 from rest_framework.decorators import action